Add a 'desktop-default' mua option
The option results in an 'xdg-open' command with a 'mailto' URI
argument, crafted from the reportbug-generated message.

Note that this drops headers that are not "To", "Subject", "Cc",
or "Bcc". In particular, "X-Debbugs-Cc" is not preserved. That
would need to be moved to a pseudo-header. I dropped a stopgap
measure that moved X-Debug-Cc entries to Cc.